Add opcode source definition for 00E0 Instruction
This commit includes the implementation of the opcode 00E0, which was
previously declared in the header file. The 00E0 instruction clears the
display, and its definition adheres to the CHIP-8 specifications.

+++ include/Chip8.hpp
@@ -88,7 +88,7 @@ private:
 
 private:
         // Registers
-        uint16_t registers[NUMBER_OF_REGISTERS];
+        uint8_t registers[NUMBER_OF_REGISTERS];
         uint16_t index_register;
         uint8_t stack_pointer;
                 // We only really need 6 bits out of this 8 bit register.
